requirements.txt

requests>=2.32.2
mwparserfromhell>=0.6.6
pydantic>=2.8.2
PyYAML>=6.0.2
tqdm>=4.66.4
tenacity>=8.4.2
opensearch-py>=2.6.0
qdrant-client>=1.9.2
neo4j>=5.23.0
beautifulsoup4>=4.12.3
html5lib>=1.1
fastapi>=0.115.0
uvicorn>=0.30.6
python-dotenv>=1.0.1
openai>=1.51.0
cohere>=5.6.2
sentence-transformers>=3.0.1
torch>=2.3.1
pandas>=2.2.2
